Add protocol support for streaming results
This may not be the best API but I am giving it a shot. Streaming
results return an iterable of defereds where they last may callback with
a None. The thought is that can be used with inlineCallbacks and be
extended to support async for without the need for the extra None.

The alternative is for users to use the protocol like interface of
message_received and connection_lost.